Sprint 18 - Stage 3 - feature_subscriptions - feature update
