--- ewm-service/src/main/java/ru.practicum.explorewithme/event/controller/AdminEventController.java
@@ -1,60 +1,0 @@
-package ru.practicum.explorewithme.event.controller; //3 - Admin endpoints
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.explorewithme.StatisticClient;
-import ru.practicum.explorewithme.event.dto.*;
-import ru.practicum.explorewithme.event.service.EventService;
-
-import javax.validation.Valid;
-import javax.validation.constraints.Positive;
-import javax.validation.constraints.PositiveOrZero;
-import java.time.LocalDateTime;
-import java.util.List;
-
-import static ru.practicum.explorewithme.constant.Constant.TIME_FORMAT;
-
-@Validated
-@RestController
-@RequiredArgsConstructor
-public class AdminEventController {
-
-    private final EventService eventService;
-    private final StatisticClient statisticClient;
-
-    @GetMapping("/admin/events")
-    public List<EventFullDto> findEventsByAdmin(@RequestParam(required = false) List<Long> users,
-                                                @RequestParam(required = false) List<String> states,
-                                                @RequestParam(required = false) List<Long> categories,
-                                                @RequestParam(required = false) @DateTimeFormat(pattern = TIME_FORMAT) LocalDateTime rangeStart,
-                                                @RequestParam(required = false) @DateTimeFormat(pattern = TIME_FORMAT) LocalDateTime rangeEnd,
-                                                @RequestParam(required = false, defaultValue = "0") Integer from,
-                                                @RequestParam(required = false, defaultValue = "10") Integer size) {
-        EventAdminParam eventAdminParam = new EventAdminParam(users, states, categories, rangeStart, rangeEnd, from, size);
-        return eventService.findEventsByAdmin(eventAdminParam);
-    }
-
-    @PatchMapping(value = "/admin/events/{eventId}")
-    public EventFullDto adminUpdateEvent(@PathVariable Long eventId,
-                                         @Valid @RequestBody UpdateEventAdminRequest updateRequest) {
-        return eventService.adminUpdateEvent(eventId, updateRequest);
-    }
-
-    @GetMapping("/users/{userId}/followers/{followerId}/events")
-    public List<EventFullDto> findEventsBySubscriptionOfUser(@PathVariable Long userId,
-                                                             @PathVariable Long followerId,
-                                                             @PositiveOrZero @RequestParam(required = false, defaultValue = "0") Integer from,
-                                                             @Positive @RequestParam(required = false, defaultValue = "10") Integer size) {
-        return eventService.findEventsBySubscriptionOfUser(userId, followerId, from, size);
-    }
-
-    @GetMapping("/users/followers/{followerId}/events")
-    public List<EventShortDto> findEventsByAllSubscriptions(@PathVariable Long followerId,
-                                                            @RequestParam(required = false, defaultValue = "NEW") String sort,
-                                                            @PositiveOrZero @RequestParam(required = false, defaultValue = "0") Integer from,
-                                                            @Positive @RequestParam(required = false, defaultValue = "10") Integer size) {
-        return eventService.findEventsByAllSubscriptions(followerId, sort, from, size);
-    }
-}
